requirements.txt
Pillow>=10.2.0
pytesseract>=0.3.10
numpy>=1.24.0
